shill: Convert code to use the newest version of libchrome.
The biggest change is a switch from using the deprecated
Task and CallbackN mechanisms to using the new Callback
mechanism.

Note: Original CL was https://gerrit.chromium.org/gerrit/16156.
This is logically another patch to that CL, but since the
latter was already merged, and is considered closed by
Gerrit, it's necessary to create a new CL.

BUG=chromium-os:15330
TEST=Build shill and run it on a zgb with a modem. Build and
run unit tests.
CQ-DEPEND=I37628863370323d30cac493764ea28f8ffd42637

Reviewed-on: https://gerrit.chromium.org/gerrit/18030
Tested-by: Eric Shienbrood <ers@chromium.org>
Commit-Ready: Eric Shienbrood <ers@chromium.org>

+++ shill/ip_address.h
@@ -44,7 +44,7 @@ class IPAddress {
 
   // Returns the prefix length given an address |family| and a |mask|. For
   // example, returns 24 for an IPv4 mask 255.255.255.0.
-  static int GetPrefixLengthFromMask(Family family, const std::string &mask);
+  static size_t GetPrefixLengthFromMask(Family family, const std::string &mask);
 
   // Returns the name of an address family.
   static std::string GetAddressFamilyName(Family family);
